+++ src/diff_evolution_solver/solver.cu
@@ -128,6 +128,15 @@ void CudaDiffEvolveSolver::MallocSetup(){
 
     CHECK_CUDA(cudaHostAlloc(&h_diversity, sizeof(float), cudaHostAllocDefault));
 
+    // arc len parameteration
+    CHECK_CUDA(cudaMalloc(&arclength::d_AllTraj_ArcLen, CUDA_SOLVER_POP_SIZE * ARC_LENGTH_SAMPLES * sizeof(float)));
+    CHECK_CUDA(cudaHostAlloc(&arclength::h_AllTraj_ArcLen, CUDA_SOLVER_POP_SIZE * ARC_LENGTH_SAMPLES * sizeof(float), cudaHostAllocDefault));
+    CHECK_CUDA(cudaMalloc(&arclength::d_AllTraj_TotalLen, CUDA_SOLVER_POP_SIZE * 1 * sizeof(float)));
+    CHECK_CUDA(cudaHostAlloc(&arclength::h_AllTraj_TotalLen, CUDA_SOLVER_POP_SIZE * sizeof(float), cudaHostAllocDefault));
+
+    CHECK_CUDA(cudaMalloc(&bezier_curve::d_all_curve_param, CUDA_SOLVER_POP_SIZE * 3 * BEZIER_SIZE * sizeof(float)));
+    CHECK_CUDA(cudaHostAlloc(&bezier_curve::h_all_curve_param, CUDA_SOLVER_POP_SIZE * 3 * BEZIER_SIZE * sizeof(float), cudaHostAllocDefault));
+
     if (DEBUG_PRINT_FLAG || DEBUG_FOOTSTEP){
         // printf("Debug flags enabled, allocating host memory\n");
         printf("bigE size:%d\n", footstep::row_bigE * footstep::col_bigE);
@@ -443,44 +452,37 @@ void CudaDiffEvolveSolver::Evaluation(int size, int epoch){
 
     CHECK_CUDA(cudaMemset(footstep::d_cluster_N_state, 0, sizeof(footstep::d_cluster_N_state)));
     // CHECK_CUDA(cudaMemset(footstep::d_D, 0, sizeof(footstep::d_D)));
-    DecodeParameters2State<<<CURVE_NUM_STEPS, CUDA_SOLVER_POP_SIZE, 0, cuda_utils_->streams_[0]>>>(new_cluster_data_, bezier_curve_manager_->curve_, footstep::d_cluster_N_state);
+    DecodeParameters2State<<<CURVE_NUM_STEPS, CUDA_SOLVER_POP_SIZE, 0, cuda_utils_->streams_[0]>>>(new_cluster_data_, bezier_curve_manager_->curve_, footstep::d_cluster_N_state, bezier_curve::d_all_curve_param);
 
     if(DEBUG_PRINT_FLAG || DEBUG_FOOTSTEP){
         // CHECK_CUDA(cudaMemcpy(footstep::h_cluster_param, new_cluster_data_->all_param, CUDA_SOLVER_POP_SIZE * CUDA_PARAM_MAX_SIZE * sizeof(float), cudaMemcpyDeviceToHost));
         CHECK_CUDA(cudaMemcpy(footstep::h_cluster_N_state, footstep::d_cluster_N_state, CURVE_NUM_STEPS * CUDA_SOLVER_POP_SIZE * footstep::state_dims * sizeof(float), cudaMemcpyDeviceToHost));
-        // CHECK_CUDA(cudaMemcpy(footstep::h_batch_D, footstep::d_batch_D, CUDA_SOLVER_POP_SIZE * sizeof(float *), cudaMemcpyDeviceToHost));
-        // float** results_h = (float**)malloc(footstep::row_D * sizeof(float*));
-        // for (int i = 0; i < footstep::row_D; ++i) {
-        //     results_h[i] = (float*)malloc(footstep::col_D * sizeof(float));
-            
-        //     // 将设备内存中的结果复制到主机内存
-        //     CHECK_CUDA(cudaMemcpy(results_h[i], footstep::h_batch_D[i], 
-        //             footstep::col_D * sizeof(float), cudaMemcpyDeviceToHost));
-            
-        //     // 打印结果
-        //     printf("Batch %d results:\n", i);
-        //     printf("[");
-        //     for (int j = 0; j < footstep::col_D; ++j) {
-        //         printf("%f ",results_h[i][j]);
-        //     }
-        //     printf("]");
-        //     printf("\n");
-        // }
-
-        // for (int i = 0; i < footstep::row_D; ++i) {
-        //     free(results_h[i]);
-        // }
-        // free(results_h);
+        CHECK_CUDA(cudaMemcpy(bezier_curve::h_all_curve_param, bezier_curve::d_all_curve_param, CUDA_SOLVER_POP_SIZE * 3 * BEZIER_SIZE * sizeof(float), cudaMemcpyDeviceToHost));
 
         // CHECK_CUDA(cudaMemcpy(host_evaluate_score_, evaluate_score_, CUDA_SOLVER_POP_SIZE * sizeof(float), cudaMemcpyDeviceToHost));
 
         // PrintMatrixByRow(footstep::h_cluster_param, CUDA_SOLVER_POP_SIZE , CUDA_PARAM_MAX_SIZE, "h_cluster_param");
         // PrintMatrixByRow(footstep::h_cluster_N_state, CUDA_SOLVER_POP_SIZE , (footstep::N+1) * footstep::state_dims, "cluster_N_state");
         PrintMatrixByRow(footstep::h_cluster_N_state, CUDA_SOLVER_POP_SIZE, CURVE_NUM_STEPS * footstep::state_dims, "before cluster_N_state");
+
+        PrintMatrixByRow(bezier_curve::h_all_curve_param, CUDA_SOLVER_POP_SIZE, 3 * BEZIER_SIZE, "all individual curve param");
         // PrintMatrixByCol(footstep::h_D, CUDA_SOLVER_POP_SIZE, footstep::N * footstep::state_dims, "before cluster_N_state");
         // PrintBatchData(footstep::h_batch_D, CUDA_SOLVER_POP_SIZE, footstep::col_D, "print batch D");
         // PrintMatrixByRow(host_evaluate_score_, CUDA_SOLVER_POP_SIZE , 1, "evaluation score");
     }
+
+    arclength::initAllTrajArcLengthMap<<<size, ARC_LENGTH_SAMPLES, 0, cuda_utils_->streams_[0]>>>(bezier_curve_manager_->curve_,
+                                                                                                 bezier_curve::d_all_curve_param, 
+                                                                                                 arclength::d_AllTraj_ArcLen,
+                                                                                                 arclength::d_AllTraj_TotalLen);
+
+    // arclength::calculateTotalLength<<<1, size, 0, cuda_utils_->streams_[0]>>>(bezier_curve_manager_->curve_,
+    //                                                                         bezier_curve::d_all_curve_param, 
+    //                                                                         arclength::d_AllTraj_TotalLen);
+    
+    CHECK_CUDA(cudaStreamSynchronize(cuda_utils_->streams_[0]));
+    CHECK_CUDA(cudaMemcpy(arclength::h_AllTraj_TotalLen, arclength::d_AllTraj_TotalLen, CUDA_SOLVER_POP_SIZE * sizeof(float), cudaMemcpyDeviceToHost));
+    PrintMatrixByRow(arclength::h_AllTraj_TotalLen, CUDA_SOLVER_POP_SIZE, 1, "all individual Traj length");
 
     CHECK_CUDA(cudaMemset(evaluate_score_, 0.0f, CUDA_SOLVER_POP_SIZE * sizeof(float)));
     footstep::EvaluatePosition<CUDA_SOLVER_POP_SIZE><<<size, footstep::N, 0, cuda_utils_->streams_[0]>>>(footstep::d_cluster_N_state, evaluate_score_);
@@ -808,6 +810,9 @@ void CudaDiffEvolveSolver::InitSolver(int gpu_device){
 
     CHECK_CUDA(cudaMemcpy(footstep::d_E_col, footstep::h_E, footstep::row_E * footstep::col_E * sizeof(float), cudaMemcpyHostToDevice));
 
+    CHECK_CUDA(cudaMemset(arclength::d_AllTraj_ArcLen, 0.0f, CUDA_SOLVER_POP_SIZE * ARC_LENGTH_SAMPLES * sizeof(float)));
+    CHECK_CUDA(cudaMemset(arclength::d_AllTraj_TotalLen, 0.0f, CUDA_SOLVER_POP_SIZE * sizeof(float)));
+
     // footstep::SetupCUDSSBatch();
     
     // magma_int_t ldda = footstep::row_bigF;
@@ -867,7 +872,7 @@ void CudaDiffEvolveSolver::InitSolver(int gpu_device){
     host_evolve_data_->problem_param.dims = dims_;
     host_evolve_data_->problem_param.int_var_dims = int_var_dims_;
 
-    host_evolve_data_->problem_param.max_round = 100000;
+    host_evolve_data_->problem_param.max_round = 0;
 
     host_evolve_data_->problem_param.accuracy_rng = 0.5;
 
@@ -1041,7 +1046,7 @@ CudaParamIndividual CudaDiffEvolveSolver::Solver(){
     //     // CHECK_CUDA(cudaMemcpyAsync(host_evolve_data_, evolve_data_, sizeof(CudaEvolveData), cudaMemcpyDeviceToHost, cuda_utils_->streams_[0]));
     //     // printf("CUDA_MAX_FLOAT %f\n", CUDA_MAX_FLOAT);
     // }
-    DecodeParameters2State<CUDA_SOLVER_POP_SIZE*3><<<1, CURVE_NUM_STEPS, 0, cuda_utils_->streams_[0]>>>(old_cluster_data_, bezier_curve_manager_->curve_, footstep::d_sol_state, true);
+    DecodeParameters2State<CUDA_SOLVER_POP_SIZE*3><<<1, CURVE_NUM_STEPS, 0, cuda_utils_->streams_[0]>>>(old_cluster_data_, bezier_curve_manager_->curve_, footstep::d_sol_state, bezier_curve::d_all_curve_param, true);
     
     const size_t gemm_shared_mem_size = footstep::F_invD_col().shared_memory_size;
     dim3 dim = footstep::F_invD_col().block_dim;